deleted old tilemap and refactored the packages to be more consice
