+++ src/game/entity/Entity.java
@@ -1,8 +1,7 @@
 package game.entity;
 
-import game.graphics.Animated;
-import game.graphics.AnimationSpriteSheet;
-import game.graphics.SpriteSheet;
+import game.graphics.animation.Animated;
+import game.graphics.sheets.AnimationSpriteSheet;
 
 public abstract class Entity extends Animated {
