+++ src/game/entity/Player.java
@@ -1,6 +1,6 @@
 package game.entity;
 
-import game.graphics.AnimationSpriteSheet;
+import game.graphics.sheets.AnimationSpriteSheet;
 
 public class Player extends Entity {
     public Player(AnimationSpriteSheet spriteSheet) {
