--- src/game/graphics/tiles/TileMap.java
@@ -1,7 +1,0 @@
-package game.graphics.tiles;
-
-import java.awt.*;
-
-public abstract class TileMap {
-    public abstract void render(Graphics2D graphics2D);
-}
